Add sanitization in addition to validation to ensure expected PHP types

+++ plugins/optimization-detective/tests/storage/test-class-od-url-metrics-post-type.php
@@ -94,7 +94,7 @@ class Test_OD_Storage_Post_Type extends WP_UnitTestCase {
 					'width'  => 640,
 					'height' => 480,
 				),
-				'timestamp' => (int) microtime( true ), // Integer to facilitate equality tests.
+				'timestamp' => microtime( true ),
 				'elements'  => array(),
 			),
 		);
